changed lots of ints to bytes

+++ Arduino/libraries/CommandProcessor/command_processor.h
@@ -12,12 +12,12 @@ class CommandProcessor
   public:
   char str[MAX_STRING_LENGTH];
   char arg[MAX_STRING_LENGTH];
-  int sub_args[NUM_SUB_ARGS];
+  byte sub_args[NUM_SUB_ARGS];
   char **commands; 
-  int num_commands;
+  byte num_commands;
   HardwareSerial *serial;
 
-  void begin(HardwareSerial *serial, char *commands[], int num_commands);
+  void begin(HardwareSerial *serial, char *commands[], byte num_commands);
   bool input_available();
   bool received_command();
   void acknowledge_command();
@@ -32,7 +32,7 @@ class CommandProcessor
   int get_sub_args();
 };
 
-void CommandProcessor::begin(HardwareSerial *serial, char **commands, int num_commands){
+void CommandProcessor::begin(HardwareSerial *serial, char **commands, byte num_commands){
   this->serial = serial;
   this->commands = commands;
   this->num_commands = num_commands;
@@ -102,7 +102,7 @@ void CommandProcessor::reset_args(){
 }
 
 int CommandProcessor::get_command(){
-  for(int i = 0; i < num_commands; i++){
+  for(byte i = 0; i < num_commands; i++){
     if(is_command(str, commands[i])){
       return CMD_FIRST + i;
     }
